+++ backend/src/main/java/com/utn/magtea/suero/SueroCreateDTO.java
@@ -6,6 +6,7 @@ import jakarta.validation.constraints.NotEmpty;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.PositiveOrZero;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.List;
 
@@ -16,5 +17,5 @@ public record SueroCreateDTO(
         @NotEmpty(message = "Debe incluir al menos un tubo")
         @Valid List<TuboInputDTO> tubos,
         @NotNull(message = "La fecha de extracción es obligatoria") LocalDate fechaExtraccion,
-        @PositiveOrZero(message = "El valor de anticuerpos no puede ser negativo") Double valorAnticuerpos
+        @PositiveOrZero(message = "El valor de anticuerpos no puede ser negativo") BigDecimal valorAnticuerpos
 ) {}
