+++ backend/src/main/java/com/utn/magtea/suero/SueroDisponibilidadDTO.java
@@ -1,9 +1,11 @@
 package com.utn.magtea.suero;
 
+import java.math.BigDecimal;
+
 public record SueroDisponibilidadDTO(
         SueroUso uso,
         Integer rango,
         Long cantidadSueros,
-        Double mlDisponibles,
+        BigDecimal mlDisponibles,
         Integer ratonesPosibles
 ) {}
