+++ backend/src/main/java/com/utn/magtea/suero/SueroResponseDTO.java
@@ -2,6 +2,7 @@ package com.utn.magtea.suero;
 
 import com.utn.magtea.tubo.TuboDTO;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.List;
@@ -16,9 +17,9 @@ public record SueroResponseDTO(
         Integer cajaNumero,
         List<TuboDTO> tubos,
         LocalDate fechaExtraccion,
-        Double cantidadTotal,
-        Double cantidadRestante,
-        Double valorAnticuerpos,
+        BigDecimal cantidadTotal,
+        BigDecimal cantidadRestante,
+        BigDecimal valorAnticuerpos,
         Integer rango,
         SueroUso uso,
         boolean activo,
