+++ backend/src/main/java/com/utn/magtea/suero/SueroUpdateDTO.java
@@ -6,6 +6,7 @@ import jakarta.validation.constraints.NotEmpty;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.PositiveOrZero;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.List;
 
@@ -15,5 +16,5 @@ public record SueroUpdateDTO(
         @NotEmpty(message = "Debe incluir al menos un tubo")
         @Valid List<TuboInputDTO> tubos,
         @NotNull(message = "La fecha de extracción es obligatoria") LocalDate fechaExtraccion,
-        @PositiveOrZero(message = "El valor de anticuerpos no puede ser negativo") Double valorAnticuerpos
+        @PositiveOrZero(message = "El valor de anticuerpos no puede ser negativo") BigDecimal valorAnticuerpos
 ) {}
