+++ backend/src/main/java/com/utn/magtea/tubo/TuboInputDTO.java
@@ -4,8 +4,10 @@ import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
 
+import java.math.BigDecimal;
+
 public record TuboInputDTO(
         @NotBlank(message = "La posición del tubo es obligatoria") String posicion,
         @NotNull(message = "La cantidad inicial es obligatoria")
-        @Positive(message = "La cantidad inicial debe ser mayor a 0") Double cantidadInicial
+        @Positive(message = "La cantidad inicial debe ser mayor a 0") BigDecimal cantidadInicial
 ) {}
